Chat, refactoring.
- Added chatbar, chatbubbles, handlers and packets.
- Re-added jumpdown.
- Clean-up and refactoring of net stuff.

+++ Graphics/GraphicsD2D.h
@@ -39,9 +39,9 @@ namespace Graphics
 		bool available(size_t id);
 		void draw(size_t, int16_t, int16_t, int16_t, int16_t, float, float, float, int16_t, int16_t);
 
-		vector2d<uint16_t> createlayout(const wstring&, Textlabel::Font, uint16_t, map<size_t, float>&);
+		vector2d<float> createlayout(const wstring&, Textlabel::Font, float, map<size_t, float>&);
 		void drawtext(const wstring&, Textlabel::Font, Textlabel::Textcolor, 
-			Textlabel::Background, float, vector2d<float>, vector2d<uint16_t>);
+			Textlabel::Background, float, vector2d<float>, vector2d<float>);
 	};
 }
 #endif
