+++ IO/Components/Nametag.cpp
@@ -24,7 +24,7 @@ namespace IO
 	{
 		if (tag.count(selected))
 		{
-			int16_t width = name.getdimensions().x();
+			int16_t width = static_cast<int16_t>(name.getdimensions().x());
 			vector2d<int16_t> startpos = position - vector2d<int16_t>(8 + width / 2, -3);
 
 			using::Graphics::DrawArgument;
